Remove HashRouter from root component.
At the present time, react-router will not be incorporated into the core
functionality. This motiviation for this decision is to minimize
reliance on external packages. Additionally, the current design
philosophy emphasizes generalizing as many features as possible to
maximize flexibility for any potential application, including those
which do not intend to utilize react-router.